:sparkles: AccountAdderForm[Component] into addAccount[Page]

+++ src/pages/api/v1/accounts/remove.js
@@ -8,9 +8,18 @@ const removeAccountApi = async (req, res, session) => {
 
             try {
                 const database = await openDatabase();
-                await database.all(`delete from Accounts where id='${data.id}'`);
+                const deletedAccount = await database.all(`select role from Accounts where id=${data.id}`);
+                const deletedRole = deletedAccount[0].role;
+                const deleterRole = session.role;
 
-                res.status(200).end();
+                const levelRole = { admin: 3, manager: 2, employee: 1 };
+                const canDelete = levelRole[deleterRole] > levelRole[deletedRole];
+                if (canDelete) {
+                    await database.all(`delete from Accounts where id='${data.id}'`);
+                    res.status(200).end();
+                } else {
+                    res.status(403).send({ message: 'No Permission!' });
+                }
             } catch (error) {
                 res.status(500).send({ message: error.message });
             }
